+++ R/paired.R
@@ -289,6 +289,9 @@ paired <- function(formula, data, id, na.action, subset=NULL, strata, control = 
         statList <- list()
         for(statfun2 in currstats) {
           statfun <- get_stat_function(statfun2)
+          tmp <- get0(statfun, mode = "function")
+          statfun <- if(is.null(tmp)) get(statfun, parent.frame(), mode = "function") else tmp
+
           bystatlist <- list()
           if(statfun %in% c("countrowpct", "countcellpct", "rowbinomCI", "Npct"))
           {