+++ R/tableby.R
@@ -347,6 +347,9 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, strata,
         statList <- list()
         for(statfun2 in currstats) {
           statfun <- get_stat_function(statfun2)
+          tmp <- get0(statfun, mode = "function")
+          statfun <- if(is.null(tmp)) get(statfun, parent.frame(), mode = "function") else tmp
+
           bystatlist <- list()
           if(statfun2 %in% c("countrowpct", "countcellpct", "rowbinomCI", "Npct"))
           {